added Authentication and Authorization to the AuctionsController.cs

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -42,12 +43,14 @@ public class AuctionsController(
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: add current user as seller
-        auction.Seller = "test";
+
+        if (User.Identity != null) auction.Seller = User.Identity.Name;
+        
         await _context.Auctions.AddAsync(auction);
         
         // Auction Created and Published in Service bus here
@@ -64,14 +67,18 @@ public class AuctionsController(
             newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
         var auction = await _context.Auctions.Include(x => x.Item)
             .FirstOrDefaultAsync(x => x.Id == id);
+        
+        if (auction?.Seller != User?.Identity?.Name) return Forbid();
+        
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+     
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -87,10 +94,14 @@ public class AuctionsController(
         return BadRequest("Problem saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await _context.Auctions.FindAsync(id);
+        
+        if (auction?.Seller != User?.Identity?.Name) return Forbid();
+        
         if (auction == null) return NotFound();
         _context.Auctions.Remove(auction);
